renaming project to Code Triage
emails coming from issuetriage.herokuapp.com are being flagged as spam by gmail and others since anyone can spin up an app on *.herokuapp.com and send email. Rather than hope that our subdomain will somehow be whitelisted, it is easier to use a proper domain. Some people have asked for features beyond just issues which might be added in the future, so the name code triage is more focused on fixing code and not 100% issue centric. For now I will keep the project at the same location/name if we move to an org then I might consider changing the name. For now it might be too confusing. 

issuetriage.herokuapp.com still works but codetriage.com is prefered.