Group triage emails together
This branch adds a new UserMailer method that accepts multiple issues and groups them into one single email. So if a user subscribes to 20 emails, they will get 1 email a day with 20 issues instead of 20 issues a day with 1 issue.

For now the older subscription notification code will not run and all users will receive a "digest" email once a day instead. In the future it would be trivial to add a flag to user accounts to allow them to opt to receive multiple emails instead.