Select the oldest, not random, issue for triaging from a subscribed repo
GitHub displays all the latest stuff first. As a result, issues that
haven't been updated for a while get to the bottom of the list, to the second
and consequent pages where they rot without further attention.

The #sample method will occasionally fetch one of these. Why not all?
A single comment will move a forgotten issue to the top of the list
as GitHub shows it where the issue will get the regular crowd's attention.
After that it will either be quickly closed or fixed. At the end of the
day subscribed repos will have much less rot in them.