[close #247] Check auth status
Once a day we can check a user's auth status and set their token to nil if it is invalid. When a user has an invalid token we need to warn them once a week.

This should help to mitigate #247.

Of the 3k+ users only around 147 had an invalid token which seems to have a disproportionately large affect on the quality and validity of CodeTriage.