Add badges
Badges allow visitors to see at a glance how many people are subscribed to your project and, the "status" of issues. For now this is a hard coded value based on the number of issues, but we could change it to a ratio or something else more clever.

I worry that larger projects would shy away from putting a red badge on their repo, but then again maybe it would encourage them to keep issue counts low. Curious on other's thoughts here. Ideally a green badge should be a point of pride.

Add a modal to the badge that gives url for easy copy