Fix test
Not sure if this is a change in the test runner or in Rails. Here's what's happening. We are storing `request.fullpath` and redirecting to it later, that is what this test is testing. Previously the `@fullpath` would also include query params, now it only shows the path:

```
@fullpath="/repos"
```

The other elements are sent in the request but they're in either 

```
action_dispatch.request.parameters"=>{"repo"=>{"name"=>"codetriage", "user_name"=>"codetriage"}, "controller"=>"repos", "action"=>"create"}
```

or 


```
"rack.request.form_hash"=>{"repo"=>{"name"=>"codetriage", "user_name"=>"codetriage"}}
```

or

```
"rack.request.form_vars"=>"repo%5Bname%5D=codetriage&repo%5Buser_name%5D=codetriage"
```

This fixes the test and does ensure that we're redirecting correctly, but it does mean that any redirects will not have data from a form post. I think this makes sense since you can't redirect to a POST, only a GET.

cc/ @prathamesh-sonpatki