DOCS DOCTOR INTEGRATION!!!
This is a 90% finished implementation of porting docs doctor to run inside of codetriage.

I'm going to be out next week, so I didn't want to do anything crazy like try to merge this in on a Friday. I've not tested the email sending, we also probably want to merge the emails to only send one single email instead of two.

This was easier than I expected, but still time consuming.

Btw if you didn't notice we are now running with SSL 🚀

Have a good weekend!