RepoBasedController#find_repo: Queue UpdateRepoInfoJob if old
This will check if the repository was updated over 6 hours ago, and queue a UpdateRepoInfoJob if so.  It returns the result of Repo.find_by_full_name either way, and only performs this check if the result is truthy.

I'm unsure about the performance drawbacks of this.  However, I do believe this is probably a decent solution.  All that happens is two checks.

This could be separated out into multiple methods if necessary.

[GitHub] This commit resolves codetriage/codetriage#489.