Change grammar 'a issues link' to 'an issues link'
This should change it from this:
```
Your contribution streak is at 1. Click on a issues link below to keep going. The goal is still to learn. Start taking notes on what you observe, how is it different today than what you saw from the last email? How is it the same?
```

To this:

```
Your contribution streak is at 1. Click on an issues link below to keep going. The goal is still to learn. Start taking notes on what you observe, how is it different today than what you saw from the last email? How is it the same?
```

A subtle change but I believe it to be a worthwhile one.