Shuffle order of issues
Right now all the issues that I get are in the same order every time. It’s a bit monotonous. I don’t want people always only clicking on the first one, or getting in the habit of scrolling past the first one etc.

In the future we can do something a bit smarter. Perhaps we can order by reverse clicked time so the repos that are most interesting are at the top.

To make a change like that I would likely want to introduce an actual A/B testing framework to see if it changes people’s behaviors (i.e. more clicks etc.).

In general Random seems to give people the impression that the algorithm is smarter than it really is. Any behavior we implement should be able to beat “random”, so this is a good baseline.