Sort the select options after converting to local time
The current implentation leaves the original order (UTC) but just
changes the labels so it appears out of order to the user. The
proposed change is not amazing code, but seem effective to order based
on the user's local time.

Advantages: quick, easy, effective
Disadvantages: not the prettiest code, breaks handlers I believe (but
there are none that I see)

I thought this might be a good time to pick a time as the default. It
seems unlikely that midnight UTC is many people's first choice, so I
figured after work-ish (7PM local) would be a reasonable default.

References:
 - jquery select sort: https://stackoverflow.com/a/7466196

Resolves #721.