Reduce issue Queries
When updating issues from GitHub we pull in multiple issues from GitHub. Previously we would then iterate over each of the issues to check to see if they exist in the database individually. With this change, we pull all of the issues (which are paginated at 25 per page?) from the database at once, so we should be seeing 25x fewer queries when updating issues.

Unfortunately the updates/writes to the database are still 1:1, we can work on that later.