Allow API Fetcher Retry on bad token
This commit gives a Fetcher the ability to pass in an optional retry parameter that will retry a request on failure. This works because random tokens are pulled for each request based off of the git_hub_bub initializer so if one token isn't good, try another for a few times before giving up.