Don't enqueue background errors in the background
The Sentry error jobs are HUGE and take up a lot of space in redis. It's fine for web jobs, but background workers are already running in the background, it takes more work to enqueue them, pop them off, and send them then it does to just send the error report.

In the future I would suggest either sending from the web dynos in a thread or first gzipping the body of the error before storing it in redis.