Get rid of N+1 queries by using upsert_all
Previously we were loading and instantiating issues for every issue pulled from GitHub. With this patch we no longer create issue objects and we've removed the N+1 query for the case where new issues are created. 

All updates and creates are now done via `upsert_all`. To accomplish this we need a unique index on [:number, :repo_id] for issues.