Remove COVID homepage
The number of COVID related projects Is relatively small and I don't think having them here is particularly driving engagement with those projects.

At the same time, I'm seeing that the overall CodeTriage user growth during the pandemic isn't the same as it was before (eyeballing, I could be wrong). 

I'm still open to exposing these repos somehow, but I think people who don't know what CodeTriage is are getting the mistaken impression it's ONLY covid related, which is not the case.

Let's revert back to the old homepage for now